modularize the game
