+++ chessBoard.py
@@ -1,0 +1,61 @@
+from pieces.pawn import Pawn
+from pieces.rook import Rook
+from pieces.knight import Knight
+from pieces.bishop import Bishop
+from pieces.queen import Queen
+from pieces.king import King
+
+# Class representing the chessboard
+class ChessBoard:
+    def __init__(self):
+        # Initializes the board with pieces
+        self.board = self.create_board()
+
+    # Creates the 8x8 chessboard and places all pieces
+    def create_board(self): 
+        board = [[None for _ in range(8)] for _ in range(8)]  # Create 8x8 grid
+        # Place white pawns at row 1 and black pawns at row 6
+        for i in range(8):
+            board[1][i] = Pawn('w', 'P')  # White pawn
+            board[6][i] = Pawn('b', 'P')  # Black pawn
+
+        # Define the placement of other pieces (Rooks, Knights, Bishops, Queen, King)
+        placement = ['R', 'N', 'B', 'Q', 'K', 'B', 'N', 'R']
+        for i in range(8):
+            if placement[i] == 'R':
+                board[0][i] = Rook('w', placement[i])  # White Rook
+                board[7][i] = Rook('b', placement[i])  # Black Rook
+            elif placement[i] == 'N':
+                board[0][i] = Knight('w', placement[i])  # White Knight
+                board[7][i] = Knight('b', placement[i])  # Black Knight
+            elif placement[i] == 'B':
+                board[0][i] = Bishop('w', placement[i])  # White Bishop
+                board[7][i] = Bishop('b', placement[i])  # Black Bishop
+            elif placement[i] == 'Q':
+                board[0][i] = Queen('w', placement[i])   # White Queen
+                board[7][i] = Queen('b', placement[i])   # Black Queen
+            elif placement[i] == 'K':
+                board[0][i] = King('w', placement[i])    # White King
+                board[7][i] = King('b', placement[i])    # Black King
+        return board
+
+    # Function to print the board in a human-readable format
+    def print_board(self):
+        for row in self.board:
+            print(' '.join([str(piece) if piece else '.' for piece in row]))  # Display pieces or '.' for empty squares
+
+    # Placeholder function for moving pieces (to be implemented)
+    def move_piece(self, start, end):
+        pass
+
+    # Placeholder function for validating moves (to be implemented)
+    def is_valid_move(self, start, end, player_color):
+        pass
+
+    # Placeholder function to check if pawn promotion is possible
+    def is_pawn_promotion_possible(self, start, end, player_color):
+        pass
+
+    # Placeholder function for asking player to promote pawn
+    def ask_for_pawn_promotion(self, start, end, player_color):
+        pass
