+++ pieces/king.py
@@ -1,0 +1,25 @@
+# Class representing a king piece
+class King:
+    def __init__(self, color, piece_type):
+        self.color = color
+        self.piece_type = piece_type
+
+    # Representation for printing the piece
+    def __repr__(self):
+        return self.piece_type if self.color == 'w' else self.piece_type.lower()
+
+    # Placeholder for validating king movement
+    def is_valid_move(self, start, end, player_color):
+        pass
+
+    # Placeholder for horizontal move validation
+    def is_horizontal_move(self, start, end, player_color):
+        pass
+
+    # Placeholder for diagonal move validation
+    def is_diagonal_move(self, board, start, end):
+        pass
+
+    # Placeholder for vertical move validation
+    def is_vertical_move(self, board, start, end):
+        pass
