+++ pieces/knight.py
@@ -1,0 +1,13 @@
+# Class representing a knight piece
+class Knight:
+    def __init__(self, color, piece_type):
+        self.color = color
+        self.piece_type = piece_type
+
+    # Representation for printing the piece
+    def __repr__(self):
+        return self.piece_type if self.color == 'w' else self.piece_type.lower()
+
+    # Placeholder for validating knight movement
+    def is_valid_move(self, board, start, end):
+        pass
