+++ pieces/queen.py
@@ -1,0 +1,25 @@
+# Class representing a queen piece
+class Queen:
+    def __init__(self, color, piece_type):
+        self.color = color
+        self.piece_type = piece_type
+
+    # Representation for printing the piece
+    def __repr__(self):
+        return self.piece_type if self.color == 'w' else self.piece_type.lower()
+
+    # Placeholder for validating queen movement
+    def is_valid_move(self, start, end, player_color):
+        pass
+
+    # Placeholder for horizontal movement check
+    def is_horizontal_move(self, board, start, end):
+        pass
+
+    # Placeholder for vertical movement check
+    def is_vertical_move(self, board, start, end):
+        pass
+
+    # Placeholder for diagonal movement check
+    def is_diagonal_move(self, board, start, end):
+        pass
